Add a BDD framework
Use cucumber as a testing framework for executing Gherkin scenarios, and
cucumber-cpp for integrating C++ step definitions in order to test C++ code
(e.g. libraries) standalone.

cucumber-cpp is a snapshot based on a rather old upstream commit
515e3980471d710e4f50ba12778284936403dc58, with three local changes:

 * Fix Clang warning about non-virtual destructor

   'CukeObject' and 'StepInfo' have virtual functions but non-virtual destructor
   [-Wnon-virtual-dtor].

 * Fix linking with against static gtest libraries

   gtest depends on pthread and therefore pthread needs to be added to the
   linker command line when linking against static gtest libraries.

 * Add some missing virtual destructors

cucumber-cpp like gmock is included in and built as part of the project in order
to make sure it is built with the same compiler flags.

cucumber is also an old version in order to stay compatible with cucumber-cpp.

Signed-off-by: Hannu Lounento <hannu.lounento@iki.fi>